DB 변경 - userIdx 데이터 타입 변경

+++ withpeace_server_springboot-main/src/main/java/com/example/demo/src/auth/model/PostLoginRes.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @AllArgsConstructor
 public class PostLoginRes {
-    private int userIdx;
+    private Long userIdx;
     private String jwt;
 }
